Fix the matcher comparing actual and expected `Competition` entities
Now that the league phase is optional, the matcher comparing competition
entities cannot assume that groups and legend values are always present.

Additionally, the previous version of the matcher only did a shallow
comparison of the contents of groups and legend, i.e. only checked if
the 0th element in expected and actual groups/legend was as expected.
Now the comparison is done on each item present, no matter how many are
there.