Move `GET /api/matches` to `GET /api/matches/grouped`
There is a need for an endpoint that lets us fetch a list of matches
whose ids are specified in a request parameter. The endpoint which
groups matches by their competition and the listing endpoint cannot be
merged into one endpoint with mutually exclusive request params,
because their return types are fundamentally different.

To goal is to have:
  * GET /api/matches/grouped?date=2024/01/01&utcOffset=+04:00
  * GET /api/matches?matchIds=some-uuid,other-uuid