Turn `Competition`'s `LeaguePhase` into an entity mapped with one-to-one
One-to-one relationship simplifies the encoding of league phase's
non-existence, replacing it with a single 'null'.

Now if we want to check if a competition has a league phase, we can
simply check if the league phase field is null. Previously we had to
encode non-existing league phases by declaring they have 0 rounds.