Fix `LocalDateTime` serialization
Without registering the date-time module, `LocalDateTime` is serialized
as a string instead of being serialized as an array of number - which
is not consistent with the way the match-service serializes the dates.

Without fixing this, the client is unable to read dates received from
the competition-service the same way it does which the dates from
match-service.