Provide `maxRounds` info on API endpoints which return `CompetitionDto`
A competition which has a league phase needs to provide additional
information about the maximum number of rounds to be played in that
phase. This information is useful when the client wants to know what
range of values is valid for
`GET /api/competitions/{competitionId}/league/rounds/{roundNumber}` http
requests. If the `roundNumber` is not in 1-{maxRounds} range, that http
request is always going to return 404 Not Found.

For competitions without a league phase, the value of `maxRounds` field
will be 0.